chore: Rename MainRightSide folder to MainProducts
