+++ src/components/pages/order/Main/Basket/BasketBody/BasketProducts.tsx
@@ -3,7 +3,7 @@ import { BASKET_MESSAGE, IMAGE_COMING_SOON } from "@/constants/product"
 import BasketCard from "./BasketCard"
 import { useOrderContext } from "@/context/OrderContext"
 import { findObjectById } from "@/utils/array"
-import { checkIfProductIsClicked } from "../../MainRightSide/Menu/helper"
+import { checkIfProductIsClicked } from "../../MainProducts/Menu/helper"
 import { TransitionGroup, CSSTransition } from "react-transition-group"
 import { basketAnimation } from "@/theme/animations"
 import { formatPrice } from "@/utils/maths"
