+++ src/components/pages/order/Main/MainProducts/Admin/tabsConfig.tsx
