+++ src/components/pages/order/Main/MainProducts/Menu/RibbonAnimated.tsx
